// NextJS/Next14Admin/types/profile.ts
export type Profile = {
    id: string;
    name: string;
    address: string;
    nationalityPT: string;
    nationalityEN: string;
    nationalityES: string;
    nationalityFR: string;
    email_1: string;
    email_2: string;
    email_3: string;
    email_4: string;
    phone_1: string;
    phone_2: string;
    phone_3: string;
    phone_4: string;
    linkedIn: string;
    web_1: string;
    web_2: string;
    web_3: string;
    web_4: string;
    web_5: string;
    web_6: string;
    img: string;
}